requirements.txt
streamlit
python-dotenv
pymongo
bcrypt
Pillow
pdf2image
google-generativeai